fix: remove acne in highly deformed spheres (#53 via #106)
* test: correct test parameter

* test: add old illumination tests

tests from chapter 8 that were removed during area light implementation

* test: enable disabled test

* test: add end to end torus lighting test

* test: add simpler thin shape test

only check one pixel which is known to currently fail

* test: add missing test declaration

* refactor: convert float variables to use double

* refactor: canvas writer uses double type

* refactor: replace floats with double in camera

* refactor: remove unnecessary cast to double

* refactor: remove unnecessary cast to double

* refactor: remove unnecessary cast to double

* refactor: reduce sphere method visibility

the methods were made public while debugging the cause of the acne

+++ src/main/java/com/BudgiePanic/rendering/io/CanvasWriter.java
@@ -60,10 +60,10 @@ public final class CanvasWriter {
      *     Builder, if the row has room for more colors.
      *     If the row has reached maximum length, the string builder is written to lines and a new SB is returned. 
      */
-    private static StringBuilder writeColor(StringBuilder builder, float color, List<String> lines) {
-        final float min = 0f;
-        final float max = 1f;
-        final float scale = 255f;
+    private static StringBuilder writeColor(StringBuilder builder, double color, List<String> lines) {
+        final double min = 0.0;
+        final double max = 1.0;
+        final double scale = 255.0;
         final int maxLineLength = 70;
         int colorDiscrete = (int) Math.ceil((scale * Math.clamp(color, min, max)));
         var strColor = String.format(" %d", colorDiscrete);
